save complete state, fix type issues, bypass digitize

+++ src/bartz/mcmcstep.py
@@ -511,7 +511,7 @@ def grow_move(X, var_tree, split_tree, max_split, p_nonterminal, sigma2, resid, 
     var_tree = var_tree.at[leaf_to_grow].set(var.astype(var_tree.dtype))
     
     split, num_available_split = choose_split(var_tree, split_tree, max_split, leaf_to_grow, key3)
-    split_tree = split_tree.at[leaf_to_grow].set(split)
+    split_tree = split_tree.at[leaf_to_grow].set(split.astype(split_tree.dtype))
 
     allowed = num_growable > 0
